# API-Azure/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Gerenciamento de Recursos em Nuvem",
    "description": "API para gerenciar recursos em nuvem utilizando Flask",
    "version": "1.0"
  },
...
  "paths": {
...
    "/azure/criar-grupo-recursos": {
      "post": {
        "summary": "Cria um Grupo de Recursos na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação do Grupo de Recursos",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome do Grupo de Recursos"
                },
                "regiao": {
                  "type": "string",
                  "description": "Região onde o Grupo de Recursos será criado"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grupo de Recursos criado com sucesso"
          },
          "500": {
            "description": "Erro ao criar Grupo de Recursos"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-conta-armazenamento": {
      "post": {
        "summary": "Cria uma Conta de Armazenamento na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da Conta de Armazenamento",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da Conta de Armazenamento"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conta de Armazenamento criada com sucesso"
          },
          "500": {
            "description": "Erro ao criar Conta de Armazenamento"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-vnet": {
      "post": {
        "summary": "Cria uma VNET na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da VNET",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da VNET"
                },
                "endereco": {
                    "type": "string",
                    "description": "Endereço da VNET (CIDR)"
                  }
                }
              }
            }
        ],
        "responses": {
          "200": {
            "description": "VNET criada com sucesso"
          },
          "500": {
            "description": "Erro ao criar VNET"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-subrede-publica": {
      "post": {
        "summary": "Cria uma Subrede Pública na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da Subrede Pública",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da Subrede Pública"
                },
                "endereco": {
                  "type": "string",
                  "description": "Endereço da Subrede Pública (CIDR)"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subrede Pública criada com sucesso"
          },
          "500": {
            "description": "Erro ao criar Subrede Pública"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-subrede-privada": {
      "post": {
        "summary": "Cria uma Subrede Privada na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da Subrede Privada",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da Subrede Privada"
                },
                "endereco": {
                  "type": "string",
                  "description": "Endereço da Subrede Privada (CIDR)"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subrede Privada criada com sucesso"
          },
          "500": {
            "description": "Erro ao criar Subrede Privada"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-grupo-seguranca-linux": {
      "post": {
        "summary": "Cria um Grupo de Segurança Linux na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação do Grupo de Segurança Linux",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome do Grupo de Segurança Linux"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grupo de Segurança Linux criado com sucesso"
          },
          "500": {
            "description": "Erro ao criar Grupo de Segurança Linux"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-grupo-seguranca-windows": {
      "post": {
        "summary": "Cria um Grupo de Segurança Windows na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação do Grupo de Segurança Windows",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome do Grupo de Segurança Windows"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grupo de Segurança Windows criado com sucesso"
          },
          "500": {
            "description": "Erro ao criar Grupo de Segurança Windows"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-interface-ip-linux": {
      "post": {
        "summary": "Cria uma Interface de IP Público Linux na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da Interface de IP Público Linux",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da Interface de IP Público Linux"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interface de IP Público Linux criada e associada com sucesso"
          },
          "500": {
            "description": "Erro ao criar Interface de IP Público Linux"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-interface-ip-windows": {
      "post": {
        "summary": "Cria uma Interface de IP Público Windows na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da Interface de IP Público Windows",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da Interface de IP Público Windows"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interface de IP Público Windows criada e associada com sucesso"
          },
          "500": {
            "description": "Erro ao criar Interface de IP Público Windows"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-maquina-virtual-linux": {
      "post": {
        "summary": "Cria uma Máquina Virtual Linux na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da Máquina Virtual Linux",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da Máquina Virtual Linux"
                },
                "usuario": {
                  "type": "string",
                  "description": "Nome do usuário da Máquina Virtual Linux"
                },
                "senha": {
                  "type": "string",
                  "description": "Senha da Máquina Virtual Linux"
                }
              },
              "required": ["nome", "usuario", "senha"]
              }
            }
        ],
        "responses": {
          "200": {
            "description": "Máquina Virtual Linux criada com sucesso"
          },
          "500": {
            "description": "Erro ao criar Máquina Virtual Linux"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-maquina-virtual-windows": {
      "post": {
        "summary": "Cria uma Máquina Virtual Windows na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação da Máquina Virtual Windows",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome da Máquina Virtual Windows"
                },
                "usuario": {
                  "type": "string",
                  "description": "Nome do usuário da Máquina Virtual Windows"
                },
                "senha": {
                  "type": "string",
                  "description": "Senha da Máquina Virtual Windows"
                }
              },
              "required": ["nome", "usuario", "senha"]
              }
            }
        ],
        "responses": {
          "200": {
            "description": "Máquina Virtual Windows criada com sucesso"
          },
          "500": {
            "description": "Erro ao criar Máquina Virtual Windows"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/criar-load-balancer": {
      "post": {
        "summary": "Cria um Load Balancer na Azure",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Informações para criação do Load Balancer",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "description": "Nome do Load Balancer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Load Balancer criado com sucesso"
          },
          "500": {
            "description": "Erro ao criar Load Balancer"
          }
        },
        "tags": ["Azure"]
      }
    },
    "/azure/destruir-recursos": {
      "post": {
        "summary": "Destroi todos os recursos na Azure",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Recursos na Azure destruídos com sucesso"
          },
          "500": {
            "description": "Erro ao destruir recursos na Azure"
          }
        },
        "tags": ["Azure"]
      }
    }
  }
}